Rename and update the test for the CodeRequestFormPageHandler
This change, as the name states, renames the file and class so that they
match the renamed, underlying class, and update the tests to match the
new functionality of the class.
